DG/requirements.txt
matplotlib==3.3.4
numpy==1.20.1
Pillow==10.2.0
torch==1.9.1
torchvision==0.10.1
